+++ src/entities/comment/api/post.ts
@@ -1,0 +1,18 @@
+export const addComment = async () => {
+  try {
+    const response = await fetch("/api/comments/add", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(newComment),
+    })
+    const data = await response.json()
+    setComments((prev) => ({
+      ...prev,
+      [data.postId]: [...(prev[data.postId] || []), data],
+    }))
+    setShowAddCommentDialog(false)
+    setNewComment({ body: "", postId: null, userId: 1 })
+  } catch (error) {
+    console.error("댓글 추가 오류:", error)
+  }
+}
