+++ src/entities/post/index.ts
@@ -1,1 +1,4 @@
+export * from "./api/delete"
+export * from "./api/get"
+export * from "./api/post"
 export * from "./type"
