+++ src/entities/user/index.ts
@@ -1,1 +1,1 @@
-export * from "./type"
+export * from "./model/types"
