+++ src/features/comments/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/AddCommentDialog"
+export * from "./ui/EditCommentDialog"
