+++ src/features/comments/model/types.ts
@@ -1,0 +1,6 @@
+export interface EditCommentDialogProps {
+  showEditCommentDialog: boolean
+  setShowEditCommentDialog: (show: boolean) => void
+  selectedComment: Comment
+  setSelectedComment: (comment: Comment) => void
+}
