+++ src/features/posts/index.ts
@@ -1,0 +1,3 @@
+export * from "./api/editPostApi"
+export * from "./ui/AddPostDialog"
+export * from "./ui/EditPostDialog"
