+++ src/features/posts/model/type.ts
@@ -1,0 +1,8 @@
+export interface PostsPaginationProps {
+  total: number
+  limit: number
+  skip: number
+  setSkip: (skip: number) => void
+  setLimit: (limit: number) => void
+  updateURL: () => void
+}
