+++ src/features/user/model/types.ts
@@ -1,0 +1,7 @@
+import { UserData } from "../../../entities/user/model/types"
+
+export type UserInfoModalProps = {
+  showUserModal: boolean
+  setShowUserModal: (show: boolean) => void
+  selectedUser: UserData | null
+}
