+++ src/widgets/CommentList/model/type.ts
@@ -1,0 +1,6 @@
+import { Comment } from "../../../entities/comment"
+
+export interface CommentListProps {
+  comments: Comment[]
+  postId: number
+}
