+++ src/widgets/PostTable/model/types.ts
@@ -1,0 +1,5 @@
+import { PostWithUser } from "../../../entities/post"
+
+export interface PostTableProps {
+  posts: PostWithUser
+}
